requirements.txt
asgiref==3.8.1
cfgv==3.4.0
distlib==0.3.9
Django==5.1.2
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
factory_boy==3.3.3
Faker==36.1.1
filelock==3.17.0
gunicorn==23.0.0
identify==2.6.7
Markdown==3.7
nodeenv==1.9.1
packaging==24.2
platformdirs==4.3.6
pre_commit==4.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.1
ruff==0.9.6
rules==3.5
sqlparse==0.5.1
tzdata==2025.1
virtualenv==20.29.2
whitenoise==6.8.2
